requirements.txt
anthropic
python-dotenv==1.0.1
pytest==8.3.2
pytest-mock==3.14.0
ruff==0.6.5
pre-commit==3.8.0
